Fix a bug with program falling when a TCP reconnect happens.
A problem was in:
After a reconnect, a RTU level has received data and sent this data to
high level eq into client. But at this time an object of pdu_t is already
destroyed (because there was a throw). And proto->rx_pdu is pointed to
unvailable place of memory.

It was fixed by setting a proto->rx_pdu to zero after a timeout happens.
Also, to guaranee a safety, I must remove any pdu's from the
try-catch context.